+++ src/main/gen/com/en_circle/slt/plugin/lisp/psi/LispToplevel.java
@@ -1,12 +1,17 @@
 // This is a generated file. Not intended for manual editing.
 package com.en_circle.slt.plugin.lisp.psi;
 
-import com.intellij.psi.PsiElement;
+import com.intellij.navigation.ItemPresentation;
+import com.intellij.psi.NavigatablePsiElement;
 import org.jetbrains.annotations.NotNull;
 
-public interface LispToplevel extends PsiElement {
+public interface LispToplevel extends NavigatablePsiElement {
 
   @NotNull
   LispSexpr getSexpr();
+
+  String getName();
+
+  ItemPresentation getPresentation();
 
 }