+++ src/main/gen/com/en_circle/slt/plugin/lisp/psi/LispVisitor.java
@@ -1,6 +1,7 @@
 // This is a generated file. Not intended for manual editing.
 package com.en_circle.slt.plugin.lisp.psi;
 
+import com.intellij.psi.NavigatablePsiElement;
 import com.intellij.psi.PsiElement;
 import com.intellij.psi.PsiElementVisitor;
 import org.jetbrains.annotations.NotNull;
@@ -96,7 +97,7 @@ public class LispVisitor extends PsiElementVisitor {
   }
 
   public void visitToplevel(@NotNull LispToplevel o) {
-    visitPsiElement(o);
+    visitNavigatablePsiElement(o);
   }
 
   public void visitVector(@NotNull LispVector o) {
@@ -109,6 +110,10 @@ public class LispVisitor extends PsiElementVisitor {
 
   public void visitNamedElement(@NotNull LispNamedElement o) {
     visitPsiElement(o);
+  }
+
+  public void visitNavigatablePsiElement(@NotNull NavigatablePsiElement o) {
+    visitElement(o);
   }
 
   public void visitPsiElement(@NotNull PsiElement o) {