+++ src/main/java/com/en_circle/slt/plugin/SltFoldingBuilder.java
@@ -1,7 +1,6 @@
 package com.en_circle.slt.plugin;
 
-import com.en_circle.slt.plugin.lisp.psi.LispList;
-import com.en_circle.slt.plugin.lisp.psi.LispSexpr;
+import com.en_circle.slt.plugin.lisp.LispParserUtil;
 import com.en_circle.slt.plugin.lisp.psi.LispToplevel;
 import com.intellij.lang.ASTNode;
 import com.intellij.lang.folding.FoldingBuilderEx;
@@ -45,32 +44,8 @@ public class SltFoldingBuilder extends FoldingBuilderEx implements DumbAware {
 
     @Override
     public @Nullable String getPlaceholderText(@NotNull ASTNode node) {
-        String text = "...";
         if (node.getPsi() instanceof LispToplevel toplevel) {
-            if (toplevel.getSexpr().getDatum() != null && toplevel.getSexpr().getDatum().getList() != null) {
-                LispList list = toplevel.getSexpr().getDatum().getList();
-                List<LispSexpr> elements = list.getSexprList();
-                if (elements.size() > 0) {
-                    LispSexpr head = elements.get(0);
-                    if (head.getDatum() != null && head.getDatum().getCompoundSymbol() != null) {
-                        String headName = head.getText();
-                        text = headName;
-
-                        if (headName.toLowerCase().startsWith("def") && elements.size() > 1) {
-                            text += " " + elements.get(1).getText();
-
-                            if (headName.equalsIgnoreCase("defmethod") && elements.size() > 2) {
-                                LispSexpr potentialArgs = elements.get(2);
-                                if (potentialArgs.getDatum() != null && potentialArgs.getDatum().getCompoundSymbol() != null) {
-                                    text += " " + potentialArgs.getText();
-                                }
-                            }
-                        }
-                    }
-                }
-
-                return text;
-            }
+            return LispParserUtil.determineTopLevelType(toplevel.getSexpr()).getShortForm();
         }
         return "...";
     }